fn-update-bonds/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

# pip install functions-framework
# functions-framework --target=process_yesterdays_charges

-i https://pypi.org/simple
cachetools==5.1.0; python_version ~= '3.7'
certifi==2022.5.18.1; python_version >= '3.6'
charset-normalizer==2.0.12; python_version >= '3'
google-api-core[grpc]==2.8.0; python_version >= '3.6'
google-auth==2.6.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
google-cloud-bigquery-storage==2.13.1; python_version >= '3.6'
google-cloud-bigquery==3.1.0
google-cloud-core==2.3.0; python_version >= '3.6'
google-cloud-storage==2.3.0
google-crc32c==1.3.0; python_version >= '3.6'
google-resumable-media==2.3.3; python_version >= '3.6'
googleapis-common-protos==1.56.1; python_version >= '3.6'
grpcio-status==1.46.1
grpcio==1.46.1; python_version >= '3.6'
idna==3.3; python_version >= '3'
numpy==1.22.3; platform_machine == 'arm64' and python_version < '3.10'
packaging==21.3; python_version >= '3.6'
pandas==1.4.2
proto-plus==1.20.4; python_version >= '3.6'
protobuf==3.20.1; python_version >= '3.7'
pyarrow==8.0.0; python_version >= '3.7'
pyasn1-modules==0.2.8
pyasn1==0.4.8
pyparsing==3.0.9; python_full_version >= '3.6.8'
pypdf2==1.27.12
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
pytz==2022.1
requests==2.27.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
rsa==4.8; python_version >= '3.6'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
urllib3==1.26.9; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'